Include advanced query system objects in docs
`Butler.query` has been public for a while, but the documentation for the `Query` object it returns was not included in the documentation.

This change pulls in the user-facing portions of the interface without any of the internal objects.  (The internal objects are sometimes referenced from constructors, but those constructors are documented as "do not use".)